backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
ffmpeg-python==0.2.0
pymongo==4.5.0
werkzeug==2.3.7
python-dotenv==1.0.0
